feat(usb_host_eth): support connect through usb ethernet chips

Closes AEG-2270

See merge request ae_group/esp-iot-solution!1265

+++ components/usb/iot_usbh_cdc/include/iot_usbh_cdc_type.h
@@ -13,7 +13,7 @@ extern "C" {
 
 typedef struct {
     uint8_t bmRequestType;
-    int32_t bNotificationCode;
+    uint8_t bNotificationCode;
     uint16_t wValue;
     uint16_t wIndex;
     uint16_t wLength;
